Use only one gen server and remove user_tokens per week (#498)
* Use only one gen server and remove user_tokens per week

* update docs

* Rename fn

* mix format
